+++ cheat/editor.py
@@ -1,4 +1,5 @@
 from __future__ import print_function
+from cheat.utils import Utils
 import subprocess
