+++ cheat/sheet.py
@@ -1,3 +1,4 @@
+from cheat.utils import Utils
 import io
 import os
 import shutil
